refactor(wip): Abstracts away on-chain record
